backend/requirements.txt
networkx>=3.1
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
openpyxl>=3.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
